Fixed some more warnings.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@26611 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/65c02core.c
@@ -1160,7 +1160,7 @@
           LOCAL_SET_OVERFLOW(((reg_a ^ tmp) & 0x80)      \
                              && ((reg_a ^ src) & 0x80)); \
       }                                                  \
-      reg_a = tmp;                                       \
+      reg_a = (BYTE)tmp;                                 \
       INC_PC(pc_inc);                                    \
   } while (0)
